+++ web/src/main/java/com/wiley/gr/ace/authorservices/web/controllers/UserLoginController.java
@@ -142,7 +142,7 @@ public class UserLoginController extends ASExceptionController {
 
             Users users = userLoginServiceDAO.getUserId(login.getEmailId());
             UserLogin user = new UserLogin();
-            user.setUserId(users.getUserId());
+            user.setUserId(users.getUserId().intValue());
             user.setFirstName(users.getFirstName());
             user.setLastName(users.getLastName());
             service.setPayload(user);
